Make Humanloop stream more robust
The previous version of this couldn't handle model-generated JSON very well, because of how we split on '}'. The new version ensures that we actually have valid JSON using a try-catch around JSON.parse, and if the parse fails (which happens when we split on a } that was actually contained inside the text output of the model, then the chunk gets accumulated into a tmp variable and we continue.